refactor: relocated types file, renamed as validators util and updated imports
