+++ src/utils/validators/index.ts
@@ -1,0 +1,5 @@
+export * from "./area";
+export * from "./history";
+export * from "./miscellaneous";
+export * from "./person";
+export * from "./usuarios";
