+++ src/utils/validators/person.ts
@@ -1,0 +1,21 @@
+import z from "zod";
+
+export const PersonSchema = z.object({
+  _id: z.string().uuid(),
+  rut: z.string(),
+  name: z.string(),
+  nameE: z.string(),
+  lastname: z.string(),
+  lastnameE: z.string(),
+  phone: z.string(),
+  insurance: z.string(),
+  address: z.string(),
+  bloodType: z.string(),
+  areaId: z.string(),
+  Rname: z.string(),
+  Rlastname: z.string(),
+  Rphone: z.string(),
+  EmergencyContact: z.string(),
+});
+
+export type TPerson = z.infer<typeof PersonSchema>;
